method next to abstact class
